requirements.txt
Flask==3.0.0
pandas==2.1.4
numpy>=1.24.0
scikit-learn==1.3.2
joblib==1.3.2
pymongo==4.6.1
flask-login==0.6.3
flask-bcrypt==1.0.1
python-dotenv==1.0.0
